Twitterabm.jl Documentation

Twitterabm.jl Documentation

Agent.jl Documentation

twitterabm.add_input — Function.

add_input(g::AbstractGraph, v::Integer, agent_list::AbstractArray, new_input_count=4)

Adds new ingoing edges to an agent. Two approaches:

In 80% of function calls neighbors of the current neighbors are selected as new candidates.

In 20% of function calls agent chooses randomly from whole network, but with regard to opinion similarity.

source

Tweet.jl Documentation

Here goes some sophisticated text about the tweet objects and functions